refactor: Reorganize repository structure for better maintainability
- Move frontend files (HTML, JS, assets) to public/ folder
- Move backend modules to src/ folder
- Move documentation to docs/ folder
- Move deployment scripts to scripts/ folder
- Move backup files to backup/ folder
- Update server.js imports to reference new src/ paths
- Update express.static to serve from public/ folder
- Update log file paths to logs/ directory

Benefits:
- Clean root directory (only essential files)
- Clear separation of concerns (frontend vs backend)
- Standard Node.js project structure
- Easier navigation and maintenance
