+++ docs/CLAUDE.md
