+++ scripts/screenshot-dashboard.js
