+++ src/model-metadata-fetcher.js
